chatbot/requirements.txt
fastapi
uvicorn
sse_starlette
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-groq
langserve
google-generativeai
groq
sentence-transformers
huggingface_hub
faiss-cpu
pypdf
PyPDF2
python-docx
python-multipart
bs4
requests
tiktoken
chromadb
python-dotenv
pydantic
